refactor(azure.ai.connections): flatten package layout + port #8358
Address PR #8357 review feedback from trangevi:

- Move internal/connections/cmd/ -> internal/cmd/ so commands live at
  the same depth as every other extension. Connection-context glue
  renamed connection_context.go to avoid colliding with the existing
  generic context.go verb. The RegisterCommands helper is inlined into
  internal/cmd/root.go (5 AddCommand calls; no helper package needed
  now that everything is in the same package).
- Move internal/connections/pkg/connections/ -> internal/pkg/connections/
  to match the convention used by other extensions.

Also ports the connection-side changes from PR #8358 (OAuth2 fields
and connector-name support) into the migrated copy. The plain
`git merge origin/main` dropped them because rename-detection didn't
follow our cross-directory migration; replayed by taking main's
connection.go / connection_test.go / raw_connection.go from the agents
copy and rewriting imports + example strings for the connections
extension.

Updates:

- internal/cmd/root.go: drop the connectioncmd helper-package import;
  inline the 5 CRUD AddCommand calls.
- internal/cmd/connection.go: imports azure.ai.connections/internal/exterrors
  (was the now-removed nested exterrors path); example strings use
  `azd ai connection` (was `azd ai agent connection`).
- AGENTS.md: package-layout description and one-way import contract
  updated to reference internal/cmd/ + internal/pkg/ (was
  internal/connections/cmd/ + internal/connections/pkg/).
- cspell.yaml: drop unused `connectioncmd` (the helper-package alias
  no longer exists).

Verified by running the local e2e smoke test
(Test_CLI_Connection_Smoke) end-to-end against a live Foundry project:
build, pack, publish, install, create, list, show, show --show-credentials,
update, post-update show, delete, post-delete show (expected 404) all
pass.
