+++ cli/azd/extensions/azure.ai.connections/internal/cmd/connection.go
@@ -13,8 +13,8 @@ import (
 	"strings"
 	"text/tabwriter"
 
-	"azure.ai.connections/internal/connections/pkg/connections"
 	"azure.ai.connections/internal/exterrors"
+	"azure.ai.connections/internal/pkg/connections"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/cognitiveservices/armcognitiveservices/v2"
 	"github.com/azure/azure-dev/cli/azd/pkg/azdext"
@@ -291,7 +291,7 @@ func (a *ConnectionCreateAction) Run(ctx context.Context) error {
 			)
 		}
 		if !hasConnector {
-			// BYO mode — required: authorization-url, token-url, client-id, client-secret.
+			// BYO mode ΓÇö required: authorization-url, token-url, client-id, client-secret.
 			// Optional: refresh-url, scopes.
 			missing := []string{}
 			if a.flags.authorizationURL == "" {
@@ -553,7 +553,7 @@ func (a *ConnectionUpdateAction) Run(ctx context.Context) error {
 	// Route to raw REST or typed SDK based on auth type
 	switch normalizedAuth {
 	case "oauth2", "user-entra-token", "project-managed-identity", "agentic-identity":
-		// Auth types that lack full ARM SDK support — update via raw REST
+		// Auth types that lack full ARM SDK support ΓÇö update via raw REST
 		err = rawCreateConnection(
 			ctx, connCtx,
 			a.flags.name,
