+++ cli/azd/extensions/azure.ai.connections/internal/pkg/connections/models_test.go
